+++ MeetingBar/Constants.swift
@@ -20,19 +20,19 @@ struct TitleTruncationRules {
 struct LinksRegex {
     let meet = try! NSRegularExpression(pattern: #"https?://meet.google.com/[a-z-]+"#)
     let hangouts = try! NSRegularExpression(pattern: #"https?://hangouts.google.com/[^\s]*"#)
-    let zoom = try! NSRegularExpression(pattern: #"https?:\/\/(?:[a-zA-Z0-9-.]+)?zoom.(?:us|com.cn)\/(?:j|my)\/[0-9a-zA-Z?=.]*"#)
+    let zoom = try! NSRegularExpression(pattern: #"https?:\/\/(?:[a-zA-Z0-9-.]+)?zoom.(?:us|com.cn)\/(?:j|my|w)\/[-a-zA-Z0-9()@:%_\+.~#?&=\/]*"#)
 
 
     /**
      * Examples:
      * - zoom native links: zoommtg://zoom.us/join?confno=92333341349&uname=Person&pwd=123456
      */
-    let zoom_native = try! NSRegularExpression(pattern: #"zoommtg://([a-z0-9-.]+)?zoom\.(us|com\.cn)/join[^\s]*"#)
+    let zoom_native = try! NSRegularExpression(pattern: #"zoommtg://([a-z0-9-.]+)?zoom\.(us|com\.cn)/join[-a-zA-Z0-9()@:%_\+.~#?&=\/]*"#)
     let teams = try! NSRegularExpression(pattern: #"https?://teams\.microsoft\.com/l/meetup-join/[a-zA-Z0-9_%\/=\-\+\.?]+"#)
     let webex = try! NSRegularExpression(pattern: #"https?://([a-z0-9-.]+)?webex\.com/[^\s]*"#)
     let jitsi = try! NSRegularExpression(pattern: #"https?://meet\.jit\.si/[^\s]*"#)
     let chime = try! NSRegularExpression(pattern: #"https?://([a-z0-9-.]+)?chime\.aws/[^\s]*"#)
-    let ringcentral = try! NSRegularExpression(pattern: #"https?://meetings\.ringcentral\.com/[^\s]*"#)
+    let ringcentral = try! NSRegularExpression(pattern: #"https?://([a-z0-9.]+)?ringcentral\.com/[^\s]*"#)
     let gotomeeting = try! NSRegularExpression(pattern: #"https?://([a-z0-9.]+)?gotomeeting\.com/[^\s]*"#)
     let gotowebinar = try! NSRegularExpression(pattern: #"https?://([a-z0-9.]+)?gotowebinar\.com/[^\s]*"#)
     let bluejeans = try! NSRegularExpression(pattern: #"https?://([a-z0-9.]+)?bluejeans\.com/[^\s]*"#)
@@ -61,6 +61,7 @@ struct LinksRegex {
     let discord = try! NSRegularExpression(pattern: #"(http|https|discord)://(www\.)?(canary\.)?discord(app)?\.([a-zA-Z]{2,})(.+)?"#)
     let blackboard_collab = try! NSRegularExpression(pattern: #"https?://us\.bbcollab\.com/[^\s]*"#)
     let coscreen = try! NSRegularExpression(pattern: #"https?://join\.coscreen\.co/[^\s]*"#)
+    let vowel = try! NSRegularExpression(pattern: #"https?://([a-z0-9.]+)?vowel\.com/#/g/[^\s]*"#)
 }
 
 enum CreateMeetingLinks {
@@ -160,6 +161,7 @@ enum MeetingServices: String, Codable, CaseIterable {
     case blackboard_collab = "Blackboard Collaborate"
     case url = "Any Link"
     case coscreen = "CoScreen"
+    case vowel = "Vowel"
     case other = "Other"
 
     var localizedValue: String {
@@ -244,6 +246,8 @@ enum MeetingServices: String, Codable, CaseIterable {
             return "constants_meeting_service_blackboard_collab".loco()
         case .coscreen:
             return "constants_meeting_service_coscreen".loco()
+        case .vowel:
+            return "constants_meeting_service_vowel".loco()
         case .other:
             return "constants_meeting_service_other".loco()
         case .url:
@@ -345,6 +349,9 @@ enum AppLanguage: String, Codable {
     case ukrainian = "ua"
     case russian = "ru"
     case croatian = "hr"
+    case german = "de"
+    case french = "fr"
+    case norwegian = "nb-NO"
 }
 
 struct Browser: Encodable, Decodable, Hashable {